Extract struct routines to module
